+++ packages/prensa/components/Article/Byline/Byline.js
@@ -5,6 +5,7 @@ import React from 'react'
 import Block from '../../Block'
 import Share from '../Share/index.tsx'
 import ManyAuthors from './Authors'
+import SectionTitle from '../SectionTitle'
 import * as S from './Byline.styled'
 
 const Byline = ({
@@ -18,13 +19,15 @@ const Byline = ({
   isExpanded,
   fbappid,
   pageUrl,
+  sectionTitle,
   share
 }) => {
-
   const shareData = get(share, 'byline', {})
   const shareColor = get(share, 'color', undefined)
   const shareSize = get(share, 'size', undefined)
   const shareEnabled = get(shareData, 'enabled', false)
+  const shareHasSectionTitle = get(shareData, 'sectionTitle', false)
+  const shareSectionTitleProps = get(shareData, 'sectionTitleProps', {})
 
   const AuthorInfo = () => {
     const hasAuthorEmail = author && author.email && author.email !== ''
@@ -100,15 +103,30 @@ const Byline = ({
             {isExpanded ? <BylineExpanded /> : <BylineReduced />}
           </S.BylineContent>
         </S.BylineContainer>
-        <Share
-          amp={amp}
-          color={shareColor}
-          enabled={shareEnabled}
-          fbappid={fbappid}
-          pageUrl={pageUrl}
-          size={shareSize}
-          {...shareData}
-        />
+        <Block
+          align='column'
+          lg={{ align: 'row', alignx: 'center' }}
+        >
+          {shareHasSectionTitle === true &&
+            <SectionTitle
+              mt={[2, 0]}
+              element='h6'
+              {...sectionTitle}
+              {...shareSectionTitleProps}
+            >
+              Compartilhar
+            </SectionTitle>
+          }
+          <Share
+            amp={amp}
+            color={shareColor}
+            enabled={shareEnabled}
+            fbappid={fbappid}
+            pageUrl={pageUrl}
+            size={shareSize}
+            {...shareData}
+          />
+        </Block>
       </S.Content>
     </S.Container>
   )