+++ packages/prensa/components/Article/SectionTitle/index.js
@@ -14,6 +14,9 @@ const SectionTitle = ({
   lineHeight,
   maxWidth,
   mb,
+  ml,
+  mr,
+  mt
 }) => {
   return (
     <Block maxWidth={maxWidth} width='100%'>
@@ -27,10 +30,16 @@ const SectionTitle = ({
         lg={{
           fontSize: fontSize[1],
           lineHeight: lineHeight[1],
-          mb: mb[1]
+          mb: mb[1],
+          ml: ml[1],
+          mr: mr[1],
+          mt: mt[1]
         }}
         mb={mb[0]}
-        mt={3}>
+        ml={ml[0]}
+        mr={mr[0]}
+        mt={mt[0]}
+      >
         {children} 
       </Typography>
     </Block>
@@ -45,6 +54,9 @@ SectionTitle.defaultProps = {
   fontWeight: 700,
   lineHeight: ['120%', '120%'],
   mb: [2, 2],
+  ml: [0, 0],
+  mr: [0, 0],
+  mt: [3, 0],
 }
 
 SectionTitle.propTypes = {
@@ -56,6 +68,9 @@ SectionTitle.propTypes = {
   lineHeight: PropTypes.array,
   maxWidth: PropTypes.string,
   mb: PropTypes.array,
+  ml: PropTypes.array,
+  mr: PropTypes.array,
+  mt: PropTypes.array
 }
 
 export default SectionTitle