+++ packages/xprog-prensa/.storybook/stories/primitives/ImageGallery.stories.tsx
@@ -4,8 +4,6 @@ import { ImageGallery } from '../../../primitives/ImageGallery/index'
 
 export default { title: 'Primitives/ImageGallery', component: ImageGallery }
 
-const RenderImage = ({ alt, src }) => <img alt={alt} src={src} width='100%' />
-
 export const Default = () => {
   const galleryList = [
     {
@@ -50,91 +48,10 @@ export const Default = () => {
     }
   ]
   return (
-    <>
-      <head>
-        <style>
-          {`
-            .carousel-root {
-              outline: none; 
-            }
-            .carousel {
-              position: relative;
-              width: 100%; 
-            }
-            .carousel * {
-              -webkit-box-sizing: border-box;
-              -moz-box-sizing: border-box;
-              box-sizing: border-box;
-            }
-            .carousel img {
-              width: 100%;
-              display: inline-block;
-              pointer-events: none; 
-            }
-            .carousel .carousel {
-              position: relative;
-            }
-            .carousel.carousel-slider {
-              position: relative;
-              margin: 0;
-              overflow: hidden;
-            }
-            .carousel .slider-wrapper {
-              overflow: hidden;
-              margin: auto;
-              width: 100%;
-              -webkit-transition: height 0.15s ease-in;
-              -moz-transition: height 0.15s ease-in;
-              -ms-transition: height 0.15s ease-in;
-              -o-transition: height 0.15s ease-in;
-              transition: height 0.15s ease-in;
-            }
-            .carousel .slider-wrapper.axis-horizontal .slider {
-              -ms-box-orient: horizontal;
-              display: -webkit-box;
-              display: -moz-box;
-              display: -ms-flexbox;
-              display: -moz-flex;
-              display: -webkit-flex;
-              display: flex;
-            }
-            .carousel .slider-wrapper.axis-horizontal .slider .slide {
-              flex-direction: column;
-              flex-flow: column;
-            }
-            .carousel .slider {
-              margin: 0;
-              padding: 0;
-              position: relative;
-              list-style: none;
-              width: 100%;
-            }
-            .carousel .slider.animated {
-              -webkit-transition: all 0.35s ease-in-out;
-              -moz-transition: all 0.35s ease-in-out;
-              -ms-transition: all 0.35s ease-in-out;
-              -o-transition: all 0.35s ease-in-out;
-              transition: all 0.35s ease-in-out;
-            }
-            .carousel .slide {
-              min-width: 100%;
-              margin: 0;
-              position: relative;
-              text-align: center;
-            }
-            .carousel .slide img {
-              width: 100%;
-              vertical-align: top;
-              border: 0;
-            }
-          `}
-        </style>
-      </head>
-      <ImageGallery
-        componentImage={RenderImage}
-        captionProps={{ enabled: true }}
-        items={galleryList}
-      />
-    </>
+    <ImageGallery
+      captionProps={{ enabled: true }}
+      showThumbs={true}
+      items={galleryList}
+    />
   )
 }