+++ pipelines/DeploySuccessNotification.js
@@ -41,4 +41,47 @@ function onDeploySuccess(pluginConfig, context) {
   return message
 }
 
-module.exports = { onDeploySuccess }
+function onPdsSysDeploySuccess(pluginConfig, context) {
+
+  //info search
+  const branchName = context && context.branch ? context.branch.name : false
+  const releaseVersion = context && context.nextRelease ? context.nextRelease.version : ''
+  const commitHead = context && context.nextRelease ? context.nextRelease.gitHead : ''
+
+  //info mount
+  const isQa = branchName === 'qa'
+  const prodMessage = `💿 Prensa System atualizado - *v${releaseVersion}*`
+  const qaMessage = `📦 Prensa System QA atualizado - *v${releaseVersion}*`
+  const notificationMessage = isQa !== false ? qaMessage : prodMessage
+  const commitUrl = `Ver mais: https://github.com/xproglabs/prensa-designsystem/commit/${commitHead}`
+  
+  //slack based block mount
+  const mainInformation = {
+    type: 'section',
+    text: {
+      type: 'mrkdwn',
+      text: notificationMessage
+    }
+  }
+  const secondInformation = {
+    type: 'context',
+    elements: [
+      {
+        type: 'mrkdwn',
+        text: commitUrl
+      }
+    ]
+  }
+  
+  const message = {
+    text: notificationMessage,
+    blocks: [
+      mainInformation,
+      secondInformation
+    ]
+  }  
+
+  return message
+}
+
+module.exports = { onDeploySuccess, onPdsSysDeploySuccess }